requirements.txt
streamlit
openai
python-dotenv
requests
Pillow

# Multi-Agent Architecture Dependencies
asyncio-extras

# OpenTelemetry Tracing Dependencies (for local agents)
opentelemetry-instrumentation-openai-v2==2.1b0
opentelemetry-sdk==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1

# Azure AI Foundry Dependencies (for Azure agents)
azure-ai-projects>=1.0.0
azure-identity>=1.25.0

# Additional utilities for agents
uuid
dataclasses